Retain NA row information
Hi,
Thank you for this awesome package!

I am proposing a minor addition to the output data in order to more easily retain which rows are NA. In my case it saves a couple small steps that would otherwise be calculating this info, and might useful for other users when analyzing the output.

Cheers.